pyproject.toml
[tool.poetry]
name = "terec"
version = "0.1.0"
description = "TEst REsult Collector"
authors = ["Tomek Lasica <tlasica@life.pl>"]
license = "MIT"
readme = "README.md"
packages = [
    {include = "terec/api", from = "bases"},
    {include = "terec/database", from = "components"},
    {include = "terec/model", from = "components"},
    {include = "terec/util",from = "bases"},
    {include = "terec/converters",from = "bases"},
    {include = "terec/ci_jenkins",from = "components"},
    {include = "terec/regression",from = "components"},
    {include = "terec/cli",from = "bases"},
    {include = "terec/auth",from = "components"},
]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115"
uvicorn = {extras = ["standard"], version = "^0.34"}
pika = "^1.3.2"
faker = "^19.10.0"
cassandra-driver = "^3.29.0"
tenacity = "^8.2.3"
flake8 = "^6.1.0"
httpx = "0.28.1"
python-jenkins = "^1.8.1"
codetiming = "^1.4.0"
typer = {extras = ["all"], version = "^0.15"}
click = "~8.1"
more-itertools = "^10.1.0"
validators = "^0.22.0"
rich = "^13.7.0"
plotille = "^5.0.0"
termcharts = "^1.1.2"
plotext = "^5.2.8"
aiohttp = {extras = ["speedups"], version = "^3.9.1"}
loguru = "^0.7.2"
scikit-learn = "^1.6.1"
python-levenshtein = "^0.27.1"
nltk = "^3.8.1"
orjson = "^3.10.18"
bcrypt = "^4.3.0"


[tool.poetry.group.dev.dependencies]
pylint = "^3.0.1"
isort = "^5.12.0"
pytest = "^8.3.5"
pytest-benchmark = "^4.0.0"
pytest-cov = "^4.1.0"
pytest-docker = "^3.2.1"
pytest-mock = "^3.12.0"
black = "^25.1.0"
ipykernel = "^6.25.2"
pre-commit = "^3.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
cli = "bases.terec.cli.main:app"

[tool.poe.tasks]
test = "poetry run pytest --keepalive --benchmark-skip --cov --cov-report=html:coverage_report --junitxml=pytest-report.xml"
benchmark = "poetry run pytest --keepalive test/bases/terec/api/test_benchmark_results_api.py"

[tool.poe.tasks.build]
shell = """
poetry build-project
pushd projects/terec_api
poetry build-project
docker build . -t terec/api:latest
popd
pushd projects/terec_cli
poetry build-project
docker build . -t terec/cli:latest
popd
"""
interpreter = "bash"

[tool.poe.tasks.start-local]
shell = "docker compose --project-directory deploy/local/ up --detach"

[tool.poe.tasks.stop-local]
shell = "docker compose --project-directory deploy/local/ down"

[tool.poe.tasks.start-astradb]
shell = "docker compose --project-directory deploy/astradb/ up --detach"

[tool.poe.tasks.stop-astradb]
shell = "docker compose --project-directory deploy/astradb/ down"

[tool.poe.tasks.start-api]
shell = """
export CASSANDRA_HOSTS=localhost
export CASSANDRA_KEYSPACE=terec
poetry run uvicorn --host 0.0.0.0 --port 8001 --log-config ./projects/terec_api/log_conf.yaml terec.api.main:app
"""
